Extract invokeMaven fn to new class
