+++ plugin-modernizer-core/src/main/java/io/jenkins/tools/pluginmodernizer/core/impl/PluginModernizer.java
@@ -2,24 +2,13 @@ package io.jenkins.tools.pluginmodernizer.core.impl;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.List;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import io.jenkins.tools.pluginmodernizer.core.config.Config;
-import org.apache.maven.shared.invoker.DefaultInvocationRequest;
-import org.apache.maven.shared.invoker.DefaultInvoker;
-import org.apache.maven.shared.invoker.InvocationRequest;
-import org.apache.maven.shared.invoker.InvocationResult;
-import org.apache.maven.shared.invoker.Invoker;
-import org.apache.maven.shared.invoker.MavenInvocationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-@SuppressFBWarnings(value = "PATH_TRAVERSAL_IN", justification = "false positive")
 public class PluginModernizer {
 
     private static final Logger LOG = LoggerFactory.getLogger(PluginModernizer.class);
@@ -28,11 +17,11 @@ public class PluginModernizer {
 
     private static final ObjectMapper objectMapper = new ObjectMapper();
 
-    private final GoalsListCreator goalsListCreator;
+    private final MavenGoalInvoker mavenGoalInvoker;
 
     public PluginModernizer(Config config) {
         this.config = config;
-        this.goalsListCreator = new GoalsListCreator(config);
+        this.mavenGoalInvoker = new MavenGoalInvoker(config);
     }
 
     public void start() {
@@ -44,12 +33,12 @@ public class PluginModernizer {
         if (subdirectories != null) {
             for (File pluginDir : subdirectories) {
                 try {
-                    List<String> goals = goalsListCreator.createGoalsList();
+                    List<String> goals = mavenGoalInvoker.createGoalsList();
                     if (goals == null) {
                         LOG.info("No active recipes.");
                         return;
                     }
-                    invokeMavenGoals(pluginDir, goals);
+                    mavenGoalInvoker.invokeMavenGoals(pluginDir, goals);
                 } catch (IOException e) {
                     LOG.error(e.getMessage(), e);
                 }
@@ -60,39 +49,6 @@ public class PluginModernizer {
         LOG.info("Plugins: {}", config.getPlugins());
         LOG.info("Recipes: {}", config.getRecipes());
         LOG.debug("Cache Path: {}", config.getCachePath());
-    }
-
-    private void invokeMavenGoals(File pluginDir, List<String> goals) {
-        Invoker invoker = new DefaultInvoker();
-        String mavenHome = config.getMavenHome();
-        if (mavenHome == null) {
-            LOG.error("Neither MAVEN_HOME nor M2_HOME environment variables are set.");
-            return;
-        }
-        try {
-            Path mavenHomePath = Paths.get(mavenHome).toRealPath();
-            if (!Files.isDirectory(mavenHomePath) || !Files.isExecutable(mavenHomePath.resolve("bin/mvn"))) {
-                LOG.error("Invalid Maven home directory. Aborting build.");
-                return;
-            }
-            invoker.setMavenHome(new File(mavenHome));
-            InvocationRequest request = new DefaultInvocationRequest();
-            request.setPomFile(new File(pluginDir, "pom.xml"));
-            request.setGoals(goals);
-            InvocationResult result = invoker.execute(request);
-            if (result.getExitCode() != 0) {
-                LOG.error("Build failed.");
-                if (result.getExecutionException() != null) {
-                    LOG.error("Execution exception occurred: ", result.getExecutionException());
-                }
-            } else {
-                LOG.info("Build succeeded.");
-            }
-        } catch (MavenInvocationException e) {
-            LOG.error("Maven invocation failed: ", e);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
     }
 
 }
